+++ packages/client.api/src/groupby/GroupByClause.test.ts
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-import type { AggregatableKeys, AggregationClause } from "@osdk/client.api";
 import { describe, it } from "vitest";
+import type { GroupByClause } from "./GroupByClause.js";
 
-export type huh = AggregatableKeys<
+export type F = GroupByClause<
   {
     metadata: any;
     objects: {
@@ -42,31 +42,6 @@ export type huh = AggregatableKeys<
   }["objects"]["Todo"]
 >;
 
-export type Q = AggregationClause<
-  {
-    metadata: any;
-    objects: {
-      Todo: {
-        type: "object";
-        apiName: "Todo";
-        primaryKeyApiName: "id";
-        primaryKeyType: "double";
-        links: {};
-        properties: {
-          text: {
-            type: "string";
-          };
-          id: {
-            type: "double";
-          };
-        };
-      };
-    };
-    actions: {};
-    queries: {};
-  }["objects"]["Todo"]
->;
-
-describe("AggregationClause", () => {
+describe("GroupByClause", () => {
   it("works", () => {});
 });
