+++ packages/client.api/src/object/ObjectDefinitions.test.ts
@@ -14,8 +14,9 @@
  * limitations under the License.
  */
 
-import type { Attachment, OsdkObjectPropertyType } from "@osdk/client.api";
 import { describe, expectTypeOf, it } from "vitest";
+import type { OsdkObjectPropertyType } from "../Definitions.js";
+import type { Attachment } from "./Attachment.js";
 
 describe("Object definitions", () => {
   it("correctly upgrades attachment types at conversion time", () => {
